made progress on compare map, restructuring code
