+++ superset-frontend/liq_viz_plugins/superset-plugin-chart-liq-thematic-maps/src/plugin/controlPanel.tsx
@@ -35,9 +35,10 @@ import {
 const sequentialSchemeRegistry = getSequentialSchemeRegistry();
 
 type Result = {
-  id: unknown;
+  id: number;
   database_name?: string;
   value?: string;
+  slice_name?: string;
 };
 
 type Data = {
@@ -576,413 +577,31 @@ const config: ControlPanelConfig = {
       ]
     },
     {
-      label: t('Custom Data'),
+      label: t('Compare Map'),
       expanded: false,
       controlSetRows: [
         [
           {
-            name: 'custom_name',
-            config: {
-              type: 'TextControl',
-              renderTrigger: false,
-              label: t('Custom layer name'),
-              default: ''
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_type',
-            config: {
-              type: 'SelectControl',
-              renderTrigger: false,
-              label: t('Data Source'),
-              description: t('Custom layer is either from a mapbox tileset URL or an existing table.'),
-              choices: [
-                ['tileset', 'Tileset'],
-                ['table', 'Table']
-              ],
-              placeholder: t('Select type') 
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_tileset',
-            config: {
-              type: 'TextControl',
-              renderTrigger: false,
-              label: t('Mapbox Tileset URL'),
-              visibility: ({ controls } : ControlState) => Boolean(controls?.custom_type.value === 'tileset')
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_database',
+            name: 'compare_chart',
             config: {
               type: 'SelectAsyncControl',
               multi: false,
-              label: t('Database'),
+              label: t('Chart'),
               dataEndpoint:
-                '/api/v1/database',
-              placeholder: t('Select database'),
-              onAsyncErrorMessage: t('Error while fetching databases'),
+                '/api/v1/chart/?q=(filters:!((col:viz_type,opr:eq,value:ext-liq-maps)))',
+              placeholder: t('Select chart'),
+              onAsyncErrorMessage: t('Error while fetching charts'),
               mutator: (data? : Data) => {
                 if (!data || !data.result) {
                   return [];
                 }
                 return data.result.map(o => ({
                   value: o.id,
-                  label: o.database_name,
+                  label: o.slice_name
                 }));
-              },
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_type.value === 'table')
-            },
-          },
-        ],
-        [
-          {
-            name: 'custom_schema',
-            config: {
-              type: 'SelectAsyncControl',
-              multi: false,
-              label: t('Schema'),
-              placeholder: t('Select schema'),
-              onAsyncErrorMessage: t('Error while fetching schemas'),
-              mapStateToProps(state : ControlPanelState) {
-                const db = state.controls.custom_database.value;
-                if (typeof db === 'number') {
-                  return {
-                    dataEndpoint: `/api/v1/database/${db}/schemas`
-                  }
-                }
-                return;
-              },
-              shouldMapStateToProps() {
-                return true
-              },
-              mutator: (data?: Data) => {
-                if (!data || !data.result) {
-                  return [];
-                }
-                return data.result.map(o => ({
-                  value: o,
-                  label: o,
-                }));
-              },
-              visibility: ({ controls } : ControlState) => Boolean(typeof controls.custom_database.value === 'number')
-            },
-          },
-        ],
-        [
-          {
-            name: 'custom_table',
-            config: {
-              type: 'SelectAsyncControl',
-              multi: false,
-              label: t('Table'),
-              placeholder: t('Select table'),
-              onAsyncErrorMessage: t('Error while fetching tables'),
-              mapStateToProps(state : ControlPanelState) {
-                const db = state.controls.custom_database.value;
-                const schema = state.controls.custom_schema.value;
-                if ((typeof db === 'number') && (typeof schema === 'string')) {
-                  return {
-                    dataEndpoint: `/api/v1/database/${db}/tables/?q=(schema_name:${schema})`
-                  }
-                }
-                return;
-              },
-              shouldMapStateToProps() {
-                return true
-              },
-              mutator: (data?: Data) => {
-                if (!data || !data.result) {
-                  return [];
-                }
-                return data.result.map(o => ({
-                  value: o.value,
-                  label: o.value,
-                }));
-              },
-              visibility: ({ controls } : ControlState) => Boolean(typeof controls.custom_schema.value === 'string')
-            },
-          },
-        ],
-        [
-          {
-            name: 'custom_geom',
-            config: {
-              type: 'SelectControl',
-              renderTrigger: false,
-              label: t('Geometry type'),
-              choices: [
-                ['point', 'Point'],
-                ['polygon', 'Polygon'],
-                ['polyline', 'Polyline'],
-                ['h3', 'H3']
-              ],
-              placeholder: t('Select geometry')
-            }
-          }
-        ],
-      ]
-    },
-    {
-      label: t('Custom Style'),
-      expanded: false,
-      controlSetRows: [
-        // Custom shape settings
-        [
-          {
-            name: 'custom_shape',
-            config: {
-              label: t('Shape'),
-              type: 'SelectControl',
-              renderTrigger: false,
-              choices: [
-                ['circle', 'Circle'],
-                ['square', 'Square'],
-                ['star', 'Star'],
-                ['pentagon', 'Pentagon'],
-                ['triangle', 'Triangle']
-              ],
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_geom.value === 'point')
-            },
-          }
-        ],
-        [
-          {
-            name: 'custom_color_attribute_check',
-            config: {
-              type: 'CheckboxControl',
-              default: false,
-              label: t('Color based on attribute?')
-            }
-          }
-        ],
-        // Custom color settings
-        [
-          {
-            name: 'custom_color_attribute',
-            config: {
-              label: t('Color Attribute'),
-              type: 'TextControl',
-              renderTrigger: false,
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_color_attribute_check === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_color',
-            config: {
-              label: t('Color'),
-              type: 'ColorPickerControl',
-              renderTrigger: true,
-              default: {
-                r: 31,
-                g: 168,
-                b: 201,
-                a: 100
-              },
-              visibility: ({ controls } : ControlState) => Boolean(!(controls.custom_color_attribute_check.value === true))
-            }
-          },
-          {
-            name: 'custom_color_scheme',
-            config: {
-              type: 'ColorSchemeControl',
-              label: t('Linear Color Scheme'),
-              choices: () =>
-                (sequentialSchemeRegistry.values() as SequentialScheme[]).map(value => [
-                  value.id,
-                  value.label,
-                ]),
-              default: sequentialSchemeRegistry.getDefaultKey(),
-              clearable: false,
-              description: '',
-              renderTrigger: true,
-              schemes: () => sequentialSchemeRegistry.getMap(),
-              isLinear: true,
-              mapStateToProps: (state : ControlPanelState) => ({
-                dashboardId: state?.form_data?.dashboardId,
-              }),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_color_attribute_check.value === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_color_breaks_mode',
-            config: {
-              type: 'SelectControl',
-              default: 'equal_count',
-              choices: [
-                ['custom', 'Custom'],
-                ['equal_count', 'Equal Count (Quantile)'],
-                ['equal_interval', 'Equal Interval'],
-                ['categorized', 'Categorized']
-              ],
-              renderTrigger: false,
-              label: t('Mode'),
-              description: t('Method used for color styling in thematic.'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_color_attribute_check.value === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_color_mode',
-            config: {
-              type: 'TextControl',
-              renderTrigger: false,
-              default: '',
-              label: t('Custom Mode'),
-              description: t('Specify a custom mode here for the number of classes, e.g. for 5 classes in custom mode would look something like 0,5,10,15,20,25 for breaks of 0-4, 5-9, 10-14, 15-19, 20+. Leave blank if specifying a mode above.'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_color_breaks_mode.value === 'custom') 
-                && Boolean(controls.custom_color_attribute_check.value === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_color_num_classes',
-            config: {
-              type: 'TextControl',
-              isInt: true,
-              default: 0,
-              validators: [legacyValidateInteger],
-              renderTrigger: false,
-              label: t('Classes'),
-              description: t('The number of breaks for the thematic'),
-              visibility: ({ controls } : ControlState) => Boolean(!(controls.custom_color_breaks_mode.value === 'categorized')) 
-                && Boolean(controls.custom_color_attribute_check.value === true)
-            },
-          },
-          {
-            name: 'custom_color_opacity',
-            config: {
-              type: 'SliderControl',
-              min: 0,
-              max: 1,
-              step: 0.1,
-              default: 0.5,
-              renderTrigger: true,
-              label: t('Opacity'),
-              description: t('Opacity of thematic'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_color_attribute_check.value === true)
-            }
-          }
-        ],
-        // Custom size settings
-        [
-          {
-            name: 'custom_size_attribute_check',
-            config: {
-              type: 'CheckboxControl',
-              default: false,
-              label: t('Size based on attribute?'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_geom.value === 'point')
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_size_attribute',
-            config: {
-              label: t('Size Attribute'),
-              type: 'TextControl',
-              renderTrigger: false,
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_size_attribute_check === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_size',
-            config: {
-              label: t('Size (px)'),
-              type: 'TextControl',
-              validators: [legacyValidateInteger],
-              default: 25,
-              renderTrigger: false,
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_geom.value === 'point'),
-              shouldMapStateToProps() {
-                return true;
-              },
-              mapStateToProps(state : ControlPanelState) {
-                if (state.controls.custom_size_attribute_check.value === true) {
-                  return {
-                    label: t('Min Size (px)')
-                  }
-                }
-                return;
               }
-            }
-          },
-          {
-            name: 'custom_size_multiplier',
-            config: {
-              label: t('Multiplier'),
-              description: t('Size increase multiplier for each increase in class'),
-              type: 'TextControl',
-              renderTrigger: false,
-              validators: [validateNumber],
-              default: 2,
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_size_attribute_check.value === true),
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_size_breaks_mode',
-            config: {
-              type: 'SelectControl',
-              default: 'equal_count',
-              choices: [
-                ['custom', 'Custom'],
-                ['equal_count', 'Equal Count (Quantile)'],
-                ['equal_interval', 'Equal Interval'],
-                ['categorized', 'Categorized']
-              ],
-              renderTrigger: false,
-              label: t('Mode'),
-              description: t('Method used for color styling in thematic.'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_size_attribute_check.value === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_size_mode',
-            config: {
-              type: 'TextControl',
-              renderTrigger: false,
-              default: '',
-              label: t('Custom Mode'),
-              description: t('Specify a custom mode here for the number of classes, e.g. for 5 classes in custom mode would look something like 0,5,10,15,20,25 for breaks of 0-4, 5-9, 10-14, 15-19, 20+. Leave blank if specifying a mode above.'),
-              visibility: ({ controls } : ControlState) => Boolean(controls.custom_size_breaks_mode.value === 'custom') 
-                && Boolean(controls.custom_size_attribute_check.value === true)
-            }
-          }
-        ],
-        [
-          {
-            name: 'custom_size_num_classes',
-            config: {
-              type: 'TextControl',
-              isInt: true,
-              default: 0,
-              validators: [legacyValidateInteger],
-              renderTrigger: false,
-              label: t('Classes'),
-              description: t('The number of breaks for the thematic'),
-              visibility: ({ controls } : ControlState) => Boolean(!(controls.custom_size_breaks_mode.value === 'categorized')) 
-                && Boolean(controls.custom_size_attribute_check.value === true)
             },
-          },
+          }
         ]
       ]
     }
