+++ superset-frontend/liq_viz_plugins/superset-plugin-chart-liq-thematic-maps/src/types.ts
@@ -82,29 +82,7 @@ export interface ThematicSchema {
 interface LiqThematicMapsCustomizeProps {
   boundary?: string;
   breaksMode?: string;
-  customColor?: RGBA;
-  customColorAttribute?: string;
-  customColorAttributeCheck?: boolean;
-  customColorBreaksMode?: string;
-  customColorMode?: string;
-  customColorNumClasses?: string;
-  customColorOpacity?: number;
-  customColorScheme?: string;
-  customDatabase?: string;
-  customGeom?: string;
-  customMode?: string;
-  customName?: string;
-  customSchema?: string;
-  customShape?: string;
-  customSize?: number;
-  customSizeAttribute?: string;
-  customSizeBreaksMode?: string;
-  customSizeMode?: string;
-  customSizeMultiplier?: number;
-  customSizeNumClasses?: number;
-  customTable?: string;
-  customtileset?: string;
-  customType?: string;
+  compareChart?: number;
   drivetimeBorderColor?: RGBA;
   drivetimeLinkedCharts?: string;
   features?: string[];
